# api/api.yaml
...
openapi: 3.0.3
info:
  title: codziennemotto.pl REST API
  description: codziennemotto.pl REST API
  version: 1.0.0
servers:
  - url: 'https'
paths:
...
  /text-set/create-new:
    description: Create new TextSet
    post:
      operationId: pl.codziennemotto.controllers.textset.TextSetController#createNewTextSetEndpoint
      tags:
        - TextSet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTextSetPayload'

      responses:
        '200':
          description: TextSet successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSet'
        '403':
          description: Forbidden - no valid authentication
  /text-set/{id}:
    description: Get TextSet by ID
    delete:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id

      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - you're not owner of target TextSet, or TextSet doesn't exist.
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id

      description: You must be authenticated as TextSet's owner or Reader.
      operationId: pl.codziennemotto.controllers.textset.TextSetController#textSetByIdEndpoint
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextSet"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - you're not owner or reader of target TextSet, or TextSet doesn't exist.
...
  /register:
    post:
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                emailAddress:
                  type: string
                username:
                  type: string
                password:
                  type: string
      tags:
        - Registration
      responses:
        '204':
          description: OK - Verification email is sent
        '400':
          description: Bad Request - See body to understand the error, potentially returned the "EmailOrUsernameIsTaken".
  /register/confirm:
    post:
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                emailAddress:
                  type: string
                code:
                  description: Verification code sent to your email
                  type: string
      tags:
        - Registration
      responses:
        '204':
          description: OK - User created
        '400':
          description: Bad Request - Something wrong, probably there's bad email or verification code.
...
components:
  schemas:
...
    TextSet:
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          nullable: true
...
    CreateTextSetPayload:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
...